+++ backend/src/app/services/vector_db/milvus_service.py
@@ -10,18 +10,10 @@ from langchain.schema import Document
 from pydantic import BaseModel, Field
 from pymilvus import DataType, MilvusClient
 
-from knowledge_table_api.core.config import Settings
-from knowledge_table_api.routing_schemas.query import (
-    Chunk,
-    Rule,
-    VectorResponse,
-)
-from knowledge_table_api.services.llm_service import (
-    LLMService,
-    decompose_query,
-    get_keywords,
-)
-from knowledge_table_api.services.vector_db.base import VectorDBService
+from app.core.config import settings
+from app.schemas.query import Chunk, Rule, VectorResponse
+from app.services.llm_service import LLMService, decompose_query, get_keywords
+from app.services.vector_db.base import VectorDBService
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -40,9 +32,8 @@ class MilvusMetadata(BaseModel, extra="forbid"):
 class MilvusService(VectorDBService):
     """The Milvus service for the vector database."""
 
-    def __init__(self, llm_service: LLMService, settings: Settings):
+    def __init__(self, llm_service: LLMService):
         self.llm_service = llm_service
-        self.settings = settings
         self.client = MilvusClient(
             uri=settings.milvus_db_uri,
             token=settings.milvus_db_token,
@@ -59,9 +50,7 @@ class MilvusService(VectorDBService):
 
     async def ensure_collection_exists(self) -> None:
         """Ensure the collection exists in the Milvus database."""
-        if not self.client.has_collection(
-            collection_name=self.settings.index_name
-        ):
+        if not self.client.has_collection(collection_name=settings.index_name):
             # Create the schema
             schema = self.client.create_schema(
                 auto_id=False,
@@ -80,7 +69,7 @@ class MilvusService(VectorDBService):
             schema.add_field(
                 field_name="vector",
                 datatype=DataType.FLOAT_VECTOR,
-                dim=self.settings.dimensions,
+                dim=settings.dimensions,
             )
 
             # Add the index
@@ -93,7 +82,7 @@ class MilvusService(VectorDBService):
 
             # Create the collection
             self.client.create_collection(
-                collection_name=self.settings.index_name,
+                collection_name=settings.index_name,
                 schema=schema,
                 index_params=index_params,
                 consistency_level=0,
@@ -107,7 +96,7 @@ class MilvusService(VectorDBService):
 
         # Upsert the vectors
         upsert_response = self.client.insert(
-            collection_name=self.settings.index_name, data=vectors
+            collection_name=settings.index_name, data=vectors
         )
 
         return {
@@ -186,7 +175,7 @@ class MilvusService(VectorDBService):
 
             # Search the colection
             query_response = self.client.search(
-                collection_name=self.settings.index_name,
+                collection_name=settings.index_name,
                 data=[embedded_query],
                 filter=f"document_id == '{document_id}'",
                 limit=40,
@@ -242,7 +231,7 @@ class MilvusService(VectorDBService):
 
                 # Query the collection
                 keyword_response = self.client.query(
-                    collection_name=self.settings.index_name,
+                    collection_name=settings.index_name,
                     filter=filter_string,
                     output_fields=[
                         "text",
@@ -354,7 +343,7 @@ class MilvusService(VectorDBService):
 
             # Query the collection
             keyword_response = self.client.query(
-                collection_name=self.settings.index_name,
+                collection_name=settings.index_name,
                 filter=filter_string,
                 output_fields=[
                     "text",
@@ -389,7 +378,7 @@ class MilvusService(VectorDBService):
 
         # Search the collection
         semantic_response = self.client.search(
-            collection_name=self.settings.index_name,
+            collection_name=settings.index_name,
             data=[embedded_query],
             filter=f'document_id == "{document_id}"',
             limit=40,
@@ -466,13 +455,13 @@ class MilvusService(VectorDBService):
     async def delete_document(self, document_id: str) -> Dict[str, str]:
         """Delete a document from the Milvus."""
         self.client.delete(
-            collection_name=self.settings.index_name,
+            collection_name=settings.index_name,
             filter=f'document_id == "{document_id}"',
         )
 
         # Confirm the deletion
         confirm_delete = self.client.query(
-            collection_name=self.settings.index_name,
+            collection_name=settings.index_name,
             filter=f'document_id == "{document_id}"',
         )